! flash strings to reduce heap usage ! storage fixes ! saving calibration ! sim eeprom load/save

+++ fw/silkopter/libs/util/storage.cpp
@@ -11,22 +11,22 @@ namespace storage
 	
 	struct Main_Header
 	{
-		Main_Header() : crc(0), count(0), end_offset(sizeof(Main_Header)) {}
-		crc_type crc;
-		size_t count;
-		size_t end_offset;
+		crc_type crc = 0;
+		size_t count = 0;
+		size_t end_offset = sizeof(Main_Header);
 	};
 
 	struct Header
 	{
-		Id id;
-		size_type size;
-		crc_type crc;
+		Id id = Id(0);
+		size_type size = 0;
+		crc_type crc = 0;
 	};	
 	
 	static const size_type MAX_DATA_SIZE = 255;
 	
 	static Main_Header s_main_header;
+	static bool s_is_initialized = false;
 
 	//static const size_t ITEM_SIZE = sizeof(Header);
 	
@@ -49,9 +49,7 @@ namespace storage
 		for (size_t i = 0; i < s_main_header.count; i++)
 		{
 			eeprom.read(header, offset);
-			auto* __restrict bytes = reinterpret_cast<uint8_t const*>(&header.crc);
-			crc = _crc16_update(crc, bytes[0]);
-			crc = _crc16_update(crc, bytes[1]);
+			update_crc(crc, header.crc);
 // 			eeprom.read(buffer, header.size, offset + sizeof(Header));
 // 			for (size_type j = 0; j < header.size; j++)
 // 			{
@@ -67,16 +65,7 @@ namespace storage
 		return crc;
 	}
 
-	static crc_type _compute_crc(uint8_t const* buffer, size_type size)
-	{
-		crc_type crc = 0;
-		for (size_type i = 0; i < size; i++)
-		{
-			crc = _crc16_update(crc, buffer[i]);
-		}
-		return crc;
-	}
-	static offset_type _find_offset_by_id(Id id)
+	static Offset _find_offset_by_id(Id id)
 	{
 		Header header;
 		auto& eeprom = board::get_eeprom();
@@ -86,13 +75,13 @@ namespace storage
 			eeprom.read(header, offset);
 			if (header.id.value == id.value)
 			{
-				return offset;
+				return Offset(offset);
 			}
 			offset += sizeof(Header);
 			offset += header.size;
 		}
 		//not found
-		return 0;
+		return Offset(0);
 	}
 
 	//////////////////////////////////////////////////////////////////////////
@@ -100,76 +89,93 @@ namespace storage
 	
 	void init()
 	{
+		if (s_is_initialized)
+		{
+			return;
+		}
+		s_is_initialized = true;
+		
 		auto& eeprom = board::get_eeprom();
 		
 		eeprom.read_advance(s_main_header, s_start_offset);
-		debug::printf("\nStorage: crc {}, count {}", s_main_header.crc, s_main_header.count);
+		debug::printf(F_STR("\nStorage: crc {}, count {}"), s_main_header.crc, s_main_header.count);
 
 		auto crc = _compute_main_crc();
 		if (crc != s_main_header.crc)
 		{
-			debug::printf("\nStorage corrupted. Formatting...");
+			debug::printf(F_STR("\nStorage corrupted. Formatting..."));
 			remove_all();
 		}
  		else
  		{
- 			debug::printf("\nCRC is ok");
+ 			debug::printf(F_STR("\nCRC is ok"));
  		}
+
+		//make sure we're never overriting the main header
+		s_main_header.end_offset = std::max(s_main_header.end_offset, sizeof(Main_Header));
 	}
 
 	crc_type get_crc()
 	{
+		ASSERT(s_is_initialized);
 		return s_main_header.crc;
 	}
 	
 	size_t get_record_count()
 	{
+		ASSERT(s_is_initialized);
 		return s_main_header.count;
 	}
 	
 	void set_record(Id id, uint8_t const* data, size_type size)
 	{
-		auto offset = find_record_offset_by_id(id);
+		ASSERT(s_is_initialized);
+		auto& eeprom = board::get_eeprom();
+		auto offset = find_record_offset_by_id(id).value;
 		if (offset != 0)
 		{
-			if (get_record_size(offset) != size)
+			if (get_record_size(Offset(offset)) != size)
 			{
 				//incompatible size - remove and add again
-				remove_record(offset);
+				remove_record(Offset(offset));
 				set_record(id, data, size);
 				return;
 			}
-			auto crc = _compute_crc(data, size);
-			board::get_eeprom().write(crc, offset + sizeof(Id) + sizeof(size_type));
-			board::get_eeprom().write(data, size, offset + sizeof(Header));
-			return;
+			auto crc = util::compute_crc(data, size);
+			eeprom.write(crc, offset + sizeof(Id) + sizeof(size_type));
+			eeprom.write(data, size, offset + sizeof(Header));
 		}
-		
-		offset = s_main_header.end_offset;
-		
-		Header header;
-		header.id = id;
-		header.size = size;
-		header.crc = _compute_crc(data, size);
-		board::get_eeprom().write_advance(header, offset);
-		board::get_eeprom().write(data, size, offset);
-		s_main_header.count++;
-		s_main_header.end_offset = offset + size;
+		else
+		{
+			offset = s_main_header.end_offset;
+
+			Header header;
+			header.id = id;
+			header.size = size;
+			header.crc = util::compute_crc(data, size);
+			eeprom.write_advance(header, offset);
+			eeprom.write(data, size, offset);
+			s_main_header.count++;
+			s_main_header.end_offset = offset + size;
+		}
+
 		s_main_header.crc = _compute_main_crc();
 		_write_main_header();
 	}
 	
 	void remove_record(Id id)
 	{
+		ASSERT(s_is_initialized);
 		auto offset = _find_offset_by_id(id);
-		if (offset != 0)
+		if (offset.value != 0)
 		{
 			remove_record(offset);
 		}
 	}
-	void remove_record(offset_type offset)
+	void remove_record(Offset offset)
 	{
-		if (offset < s_start_offset || offset >= s_main_header.end_offset)
+		ASSERT(s_is_initialized);
+		if (offset.value < s_start_offset || offset.value >= s_main_header.end_offset)
 		{
 			ASSERT(0);
 			return;
@@ -179,8 +185,8 @@ namespace storage
 		
 		static const size_t BUFFER_SIZE = 256;
 		uint8_t buffer[BUFFER_SIZE];
-		auto dst_offset = offset;
-		auto src_offset = offset + sizeof(Header) + get_record_size(offset);
+		auto dst_offset = offset.value;
+		auto src_offset = offset.value + sizeof(Header) + get_record_size(offset);
 		auto size = std::min(BUFFER_SIZE, s_main_header.end_offset - src_offset);
 		while (size > 0)
 		{
@@ -196,10 +202,11 @@ namespace storage
 		_write_main_header();
 	}
 		
-	size_type get_record(offset_type offset, uint8_t* dst, size_type max_size)
+	size_type get_record(Offset offset, uint8_t* dst, size_type max_size)
 	{
+		ASSERT(s_is_initialized);
 		ASSERT(dst);
-		if (offset < s_start_offset || offset >= s_main_header.end_offset)
+		if (offset.value < s_start_offset || offset.value >= s_main_header.end_offset)
 		{
 			ASSERT(0);
 			return 0;
@@ -209,24 +216,26 @@ namespace storage
 		auto size = std::min(get_record_size(offset), max_size);
 		if (size)
 		{
-			eeprom.read(dst, size, offset + sizeof(Header));
+			eeprom.read(dst, size, offset.value + sizeof(Header));
 		}
 		return size;
 	}
 	size_type get_record(Id id, uint8_t* dst, size_type max_size)
 	{
+		ASSERT(s_is_initialized);
 		ASSERT(dst);
 		auto offset = _find_offset_by_id(id);
-		if (offset != 0)
+		if (offset.value != 0)
 		{
 			return get_record(offset, dst, max_size);
 		}
 		return 0;
 	}
 
-	crc_type get_record_crc(offset_type offset)
+	crc_type get_record_crc(Offset offset)
 	{
-		if (offset < s_start_offset || offset >= s_main_header.end_offset)
+		ASSERT(s_is_initialized);
+		if (offset.value < s_start_offset || offset.value >= s_main_header.end_offset)
 		{
 			ASSERT(0);
 			return 0;
@@ -234,12 +243,13 @@ namespace storage
 
 		auto& eeprom = board::get_eeprom();
 		crc_type crc;
-		eeprom.read(crc, offset + sizeof(Id) + sizeof(size_type));
+		eeprom.read(crc, offset.value + sizeof(Id) + sizeof(size_type));
 		return crc;
 	}
-	size_type get_record_size(offset_type offset)
+	size_type get_record_size(Offset offset)
 	{
-		if (offset < s_start_offset || offset >= s_main_header.end_offset)
+		ASSERT(s_is_initialized);
+		if (offset.value < s_start_offset || offset.value >= s_main_header.end_offset)
 		{
 			ASSERT(0);
 			return 0;
@@ -247,18 +257,20 @@ namespace storage
 
 		auto& eeprom = board::get_eeprom();
 		size_type size;
-		eeprom.read(size, offset + sizeof(Id));
+		eeprom.read(size, offset.value + sizeof(Id));
 		return size;
 	}
 		
-	offset_type find_record_offset_by_id(Id id)
+	Offset find_record_offset_by_id(Id id)
 	{
+		ASSERT(s_is_initialized);
 		return _find_offset_by_id(id);
 	}
 
 
 	void remove_all()
 	{
+		ASSERT(s_is_initialized);
 		s_main_header.count = 0;
 		s_main_header.crc = 0;
 		s_main_header.end_offset = sizeof(Main_Header);
